Separate retry-after testing.  Address review comments.
