+++ tests/test_api.py
@@ -8,86 +8,17 @@ import os
 import sys
 import unittest
 import requests
-from unittest import mock
 
 pkg_root = os.path.abspath(os.path.join(os.path.dirname(__file__), '..'))  # noqa
 sys.path.insert(0, pkg_root)  # noqa
 
 import dss
-from dss.error import DSSException
 from dss.util import UrlBuilder
 from dss.util.version import datetime_to_version_format
-from dss.config import Replica, DeploymentStage, Config
+from dss.config import Replica
 from tests.infra import DSSAssertMixin, DSSUploadMixin, DSSStorageMixin, TestBundle, testmode, ExpectedErrorFields
 from tests.infra.server import ThreadedLocalServer
 from tests import get_auth_header
-
-
-class TestApiErrors(unittest.TestCase, DSSAssertMixin):
-    @unittest.skipIf(DeploymentStage.IS_PROD(), "Skipping synthetic 504 test for PROD.")
-    def test_504_post_bundle_HAS_NO_retry_after_response(self):
-        self.app = ThreadedLocalServer()
-        self.app.start()
-        self.app._chalice_app._override_exptime_seconds = 15.0
-        uuid = "xxxxxxxx-xxxx-xxxx-xxxx-xxxxxxxxxxxx"
-        version = datetime_to_version_format(datetime.datetime.utcnow())
-
-        url = str(UrlBuilder().set(path=f"/v1/bundles/{uuid}/checkout")
-                  .add_query("version", version)
-                  .add_query("replica", 'aws'))
-
-        r = self.assertPostResponse(
-            url,
-            504,
-            expected_error=ExpectedErrorFields(
-                code="timed_out",
-                status=requests.codes.gateway_timeout,
-            ),
-            headers={"DSS_FAKE_504_PROBABILITY": "1.0"}
-        )
-        self.assertTrue('Retry-After' not in r.response.headers)
-        self.app.shutdown()
-
-    @unittest.skipIf(DeploymentStage.IS_PROD(), "Skipping synthetic 504 test for PROD.")
-    def test_504_get_bundle_HAS_retry_after_response(self):
-        self.app = ThreadedLocalServer()
-        self.app.start()
-        self.app._chalice_app._override_exptime_seconds = 15.0
-        uuid = "xxxxxxxx-xxxx-xxxx-xxxx-xxxxxxxxxxxx"
-        version = datetime_to_version_format(datetime.datetime.utcnow())
-
-        url = str(UrlBuilder().set(path=f"/v1/bundles/{uuid}")
-                  .add_query("version", version)
-                  .add_query("replica", 'aws'))
-
-        r = self.assertGetResponse(
-            url,
-            504,
-            expected_error=ExpectedErrorFields(
-                code="timed_out",
-                status=requests.codes.gateway_timeout,
-            ),
-            headers={"DSS_FAKE_504_PROBABILITY": "1.0"}
-        )
-        self.assertEqual(int(r.response.headers['Retry-After']), 10)
-        self.app.shutdown()
-
-    def test_502_get_bundle_HAS_retry_after_response(self):
-        with mock.patch('dss.api.bundles.get') as foo:
-            foo.side_effect = DSSException(502, 'server_error', "Internal Server Error.")
-            self.app = ThreadedLocalServer()
-            self.app.start()
-            self.app._chalice_app._override_exptime_seconds = 15.0
-            uuid = "xxxxxxxx-xxxx-xxxx-xxxx-xxxxxxxxxxxx"
-            version = datetime_to_version_format(datetime.datetime.utcnow())
-
-            url = str(UrlBuilder().set(path=f"/v1/bundles/{uuid}")
-                      .add_query("version", version)
-                      .add_query("replica", 'aws'))
-
-            r = self.assertGetResponse(url, 502, headers=get_auth_header())
-            self.assertEqual(int(r.response.headers['Retry-After']), 10)
-            self.app.shutdown()
 
 
 @testmode.integration
